Allow writing to custom PILS parameters
When the MCU reports a custom parameter with a name,
an asyn parameter will be created
(unless already done by a previous iocsh command).
This parameter will be added to the poll list, so it
can be connected to an "ai" input record.

Handling ao records was not possible so far.
When there is such an output record, it needs to be translated
from an "asyn function" into a paramIndex used towards the MCU.
Add a function paramIndexFromFunction() to do so and use it for writing.