custom parameters: Use AnalogOutputReadback and DESC field
Improve the handling of record- and asyn-paramter names:
We can use long asyn-parameter names, as they are more descriptive.
However, record names have a maximal length.
Solution:
Use the long name when "linking" a record towards the asyn paramter.
Use the short name for record names.
Use the long name for the DESC field and show it in CSS.

Changes to be committed:
    modified:   ethercatmcApp/Db/Makefile
    deleted:    ethercatmcApp/Db/ethercatmcAnalogOutput.template
    new file:   ethercatmcApp/Db/ethercatmcAnalogOutputReadback.template
    modified:   ethercatmcExApp/op/Boy/ethercatmcaxisConfig-pils-TimeEncoder.opi
    modified:   ethercatmcExApp/op/Boy/tools/ethercatmcaxisConfigCustomPara.mid
    deleted:    iocsh/ethercatmcAnalogOutput.iocsh
    new file:   iocsh/ethercatmcAnalogOutputReadback.iocsh
    modified:   test/startup/st.mcu006-PTP-B02.iocsh