AnalogInput/AnalogOutputReadback: DESC comes from st.cmd
Originally it seemed like a good idea to get the DESC field
of e.g. an anlog input record from the MCU.
However, since we have "named asyn parameters" to "map the records onto",
it is easy enough to provide a DESC field in the st.cmd (as well as EGU)
Remove the helper asy-parameters and records.

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmcAnalogInput.template
    modified:   iocsh/ethercatmcAnalogInput.iocsh
    modified:   iocsh/ethercatmcAnalogOutputReadback.iocsh