Remove obsolete special handling for parameter 192..194
Remove the obsolete handling of parameters 192..194:
Originally defined as "integer over the wire" (instead of double)
this did never work as intended.
Already 2001 this was corrected in the PLC, and they are,
even if they are boolean or integer, transferred as floating point
over the wire. And ever since then hard coded number 218..220 are
defined.

In short: Remove old stuff.