Remove asyn parameters CfgDxLM-En-RB
Origianlly it was thought that having to records with 2 asyn parameters
for each soft-limit-enable, high- and low-) was a good idea,
One for writing towards the MCU, and one for the readback.
However, since there is only one variable in the MCU anyway,
use the "info:readback" for the output record, to synchronize the
record with the reading from the MCU.

Remove the "-RB" asyn parameters, keep 2 different records:
The -RB as an input record for the default template,
and the non-RB for an additional output record in the debug template.

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmc.template
    modified:   ethercatmcApp/src/ethercatmcAxis.cpp
    modified:   ethercatmcApp/src/ethercatmcController.cpp
    modified:   ethercatmcApp/src/ethercatmcController.h
    modified:   ethercatmcApp/src/ethercatmcIndexer.cpp