Make setting of move/idle current work
Implement reading/writing of current parameters.
This must be supported by the MCU, of course.
Defrin 2 new asyn parameters to host parameter index 62 and 64,
which are defined in PILS, but had never been implemented.

Rename all asy parameters, as we now do not (mis-) use one of the custom
parameters any more.

Changes to be committed:
    modified:   ethercatmcApp/src/ethercatmcController.cpp
    modified:   ethercatmcApp/src/ethercatmcController.h
    modified:   ethercatmcApp/src/ethercatmcIndexer.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp
    modified:   ethercatmcExApp/op/Boy/ethercatmcaxisConfig-pils-Max-Red-Current-TimeEncoder.opi
    modified:   ethercatmcExApp/op/Boy/tools/Makefile
    modified:   test/startup/st.mcu006-PTP-B02.iocsh