test/simulator/sock-util.c: IPv4 only
For some testing the listening ports on the simulator
had been opened with AF_INET6.
Under Linux and MacOs thoses sockets accept IPV4 as well,
unless the socket had been bound to IPv4 already.
Then silently only IPv6 traffic is accepted.
The disadvantage here is that we fail, if e.g.
the ads-forwarder is running.

Another disadvantage is, that FreeBSD does not accept
IPv4 on such a socket, and that is what we need.

Solution: Open the socketswith AF_INET again.
When the day comes, that IPv6 is needed for the simulator,
more coding may be needed.