ethercatmcAxis.cpp: Not homed sets no longer the problem bit
"Old" versions of the for ess-motorRecord did treat the "problem" bit
not as indicator, the the motor has stopped.
Later versions do.
This makes the DMOV bit toggle twice when a not homed motor is homed.

The alarm handling (for a not homed motor) is now done via setting
motorNotHomedProblem_ to 1.

In other words: It is time to remove this handling:
eeAxisError = eeAxisErrorNotHomed