opi/Boy/ethercatmcaxisExpert-piezo.opi: Better PSTAT
First step towards an opi that allow a 32 bit word to be displayed
in the singe bits.
So far the "AUX" bits have been doing that, each bit has it's own
string record showing the name of this bit.
For the PSTAT there is no need to have run-time-dependend names.
Everything is fixed in the controller.
Prepare a way to generate a number of hard-coded text fields
via Makefile and genExpertWithAuxBits.sh to assemble
ethercatmcaxisExpert-piezo.opi with help of the new
ethercatmcaxisPiezoPSTAT.mid
TODO: Adapt the name of the PV, to be called -PSTAT, after changing
the IOC template.
Changes to be committed:
    modified:   ../ethercatmcaxisExpert-piezo.opi
    modified:   Makefile
    new file:   ethercatmcaxisPiezoPSTAT.mid
    modified:   genExpertWithAuxBits.sh