ethercatmcIndexerAxis: Do not poll parameters in localMode
The bit "localMode" can be set inside the MCU to prevent EPICS
"to send signals".
That is working as expeted, but it turned out that reading a
parameter is "sending a signal" (even if not moving a motor).
Avoid that sending as well, to allow a GUI on the MCU
to be used to read- or write parameters undisturbed.