test...indexer.c: Consistant soft limits
Make the soft limit of the simulator (more) consistent:
The end of the travelling range is where the limit switch is.
This is NOT the range where soft limits are defined inside the MCU.
They are some distance away from the "hard limit", 1mm in our case.
However, in the real MUC, the absMax and absMin values are derived
from the soft limit set by the motion control engineers, so they
have the same value.
Do the same for the similator, avoiding a "jump" from e.g. 6.0 mm
for the low limit into 5.0 mm when the test suite tries to restore
lost soft limits.

See this code in AxisMr.py:

    def setSoftLimitsOn(
        self, tc_no, low_limit=0.0, high_limit=0.0, initAbsMinMax=False
    ):
        if initAbsMinMax:
            high_limit = self.axisCom.get("-CfgPMAX-RB")
            low_limit = self.axisCom.get("-CfgPMIN-RB")