pytest36: Return values for AxisMr/jogDirectionTimeout()
Start to add checks for "DMOV" blinking feature.
Equip jogDirectionTimeout() with a return value (True/False) and use it.
Some more changes to allow waitForStartAndDone() to either throw an
exception when things fail, or just to return False.
This needs to ripple into more test cases in the next commits

Changes to be committed:
    modified:   110_Ethercat_HOMF_-ProcHom.py
    modified:   121_Record-JOGF_DIR_DLY_HLM.py
    modified:   131_Record-JOGR_DIR_DLY_LLM.py
    modified:   AxisMr.py