test/AxisMr: Clean up moveWait()
Use the functionality of the motorRecord:
When writing to the .VAL field (and others) we can wait for a callback.
Use this: Calculate the timeout and wait.

This needs some more changes, e.g. in AxisCom to allow an exeption
not to be thrown, the same feature that p4p already has.
p4p will not throw an exception, if asked to do so, but return it.
We use a different approach: put() returns True or False.
Or throws an exception. Whatever throw=throw asks for.

Changes to be committed:
    modified:   pytests36/140_Record-VAL_VMAX.py
    modified:   pytests36/141_Record_NTM.py
    modified:   pytests36/160-Ethercat-MoveAbs-STOP.py
    modified:   pytests36/AxisCom.py
    modified:   pytests36/AxisMr.py