test...AxisCom.py: Correct throw=throw for p4p
The problem:
p4p will either throw an exception (if needed and throw=True) or
return the exception as a return value (or None).
The logic here says that True or False should be returned (throw=False)
or True should be returned anyway or an exception to be thrown.

The logic for p4p did not return True, when an exception should not
be thrown and everything was OK.