Add ethercatmcAnalogInputSmooAdelMdel.template
commit 7da81da5d9f73f3694a1678c7b9379a81578dda5
    ethercatmcAnalogInput.template: Add SMOO, ADEL, MDEL
    Make it possible to use SMOO, ADEL and MDEL
was not ideal: Using ethercatmcAnalogInput.iocsh gave warnings.
E.g. the SMOO macros was undefined.

Fix this by adding a new template:
ethercatmcAnalogInput.iocsh uses ethercatmcAnalogInput.template
ethercatmcAnalogInputSmooAdelMdel.iocsh uses
ethercatmcAnalogInputSmooAdelMdel.template

Note: A different solution may be to define SMOO as 0 in
ethercatmcAnalogInput.iocsh and revert this commit.
That may be done in the future.
Changes to be committed:
    modified:   Db/Makefile
    modified:   Db/ethercatmcAnalogInput.template
    new file:   Db/ethercatmcAnalogInputSmooAdelMdel.template
    modified:   ../iocsh/ethercatmcAnalogInputSmooAdelMdel.iocsh