Support for writing HomProc and HomPos
Add 2 records to all testing of homing procedures:
"-HomProc" and "-HomPos"
Both are located in the debug template.
Note that writing these paramaters must still be allowed from
the motion controller.
Having these new records does not guarantee that we can mess with the
homing on all devices from a user perspective.
This is still reserved for testing and/or test crates done by experts.

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmcdebug.template
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp