ethercatmcIndexerV2: Fix memory overwrite for custom parameters
Fix a memory overwrite introduced in
3667d93b4f, "custom parameters are per axis"

The name of a custom parameter is now stored in a per-axis basis.
However, this commit introduced a bug:
The array was too short.
Introduce a new define "PARAM_IF_NUM_CUSTOM_PARAS" which is
corrected to be 32 instead of 31.

Rename customParaName into customParaNames to make more clear
that this is a 2-dimensional array.

And finaly, use strncpy() instead of memcpy() to make more clear
that we are dealing with strings.

Changes to be committed:
    modified:   ethercatmcApp/src/ethercatmcController.h
    modified:   ethercatmcApp/src/ethercatmcIndexer.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.h
    modified:   ethercatmcApp/src/ethercatmcIndexerV2.cpp