Add ethercatmcConfigParaDesc.template
Add a new template and iocsh to load custom parameters:
They are of type "ao" to set different parameters inside the MCU
When the parameter inside the MCU is an integer (typical int32)
it will be converted inside the MCU.
EPICS will always use a floating point (double, 64 bit)
which can hold all valid values for an int32 as well.
To describe what the parameter does there is a stingin record,
which needs to be added to the CSS screens in the next commits.

Changes to be committed:
    modified:   ethercatmcApp/Db/Makefile
    new file:   ethercatmcApp/Db/ethercatmcConfigParaDesc.template
    new file:   iocsh/ethercatmcConfigParaDesc.iocsh