test...101_Record-VAL-HOMF-VAL: ethercatmc softlimits if needed
Add a fallback when the upstream motor module is used:
The "readonly soft limits from the controller" exist in the controller and
are obeyed but do not ripple into the motorRecord.HLM and .DLM fields.
Add a workaound: when .DLM == .HLM == 0.0 try to read the
ethercatmc-specific PVs "-CfgDLLM-RB" and "-CfgDHLM-RB" and use them.
This is not portable towards other controller, but that is OK since
this test case is part of ethercatmc.