test: Fix possible failure of tc122
Fix a failure when test cases are executed against a running IOC:
- test case 121 sets the .DIR field to 1, and does not restore it:
  Restore DIR (and FOFF) after the test case.
- test case 122 fails because AxisMr::moveIntoLS() looks at the wrong
  limit switch: When moving a motor with JOGF we need to check
  for HLS. And not the raw-bit in MSTA.
  Fix this.

The whole problem was not found by ci because each test case that is run
under the ci scripts is starting a fresh IOC with .DIR == 0.

Do another cleanup:
make 121 and 131 as similar as possible in the source code.

Changes to be committed:
    modified:   test/pytests36/121_Record-JOGF_DIR_DLY_HLM.py
    modified:   test/pytests36/131_Record-JOGR_DIR_DLY_LLM.py
    modified:   test/pytests36/AxisMr.py