test...AxisMr: moveWait and postMoveCheck print improvements
Improve the printing in moveWait():
Before it could happen that ret was printed as true,
but the comparison about calbacks made thr function return False.
This was not really visable in the log.
Introduce ret0 as an intermediate value, print it together with ret
and return what was printed.

Improve postMoveCheck(): All values are casted into int or float.
Do not check lvio any more (more investigation needed in motorRecord.cc)