Remove ethercatmcIndexerV3.cpp
Remove (unused) code that had been developped but will not be used
in a forseeable future: ethercatmcIndexerV3.cpp

As it looks like, the MCU software will stay at v2 for a while.
Code that is not used everyday (or in this case, not at all)
should be removed for reasons like maintainability, readability
or to save compilation time for nothing.
The good thing is that Git allows to code to be retrieved again.

Changes to be committed:
    modified:   ethercatmcApp/src/Makefile
    modified:   ethercatmcApp/src/ethercatmcController.h
    modified:   ethercatmcApp/src/ethercatmcIndexer.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp
    deleted:    ethercatmcApp/src/ethercatmcIndexerV3.cpp