test...-one-TC.sh: Prefix printouts
The test script starting the simulator, the IOC and then
running all test collects just one log file
(That is what a ci script on a foreign git server allows us)
In that sense the printouts from the 3 processes are all
bundled together.

Mark printouts from pytest with "PYT:" and the simulator with "SIM:"
Trying to mark printouts from the IOC with IOC: didn't work
as expected. Suspected to be caused by buffering in e.g. asynPrint ?

So let's leave the IOC log unmarked.

Changed file:
test/pytests36/test-ioc-with-sim-indexer-one-TC.sh