test...AxisMr: correct setMotorStartPos
Use the correct method towards the simulator:
fActPosition will "home" the motor. When the position is wrong,
we may run into the limit switches at unexpected positions.
Use the correct fSimForcePos instead, which is "only" doing a fast move
to "move" the motor fast to a start position without the need wait
for the real move with velocity.