ethercatmcController.h: Proper locking of the serial port
Fix a long standing design-miss:
While debugging another issue it turned out that the calls
to queueLockPort() and queueUnlockPort() had been missing when an
ADS transaction is made.
The code should go like this:
- queueLockPort()
- send request
- read first part of response (AMS header)
   Now we have the length of the response
 - read second part of response (if any)
- queueUnockPort()

That had been not a problem so far, because only one asnyport ("MCU1")
is using the communication asyn port ("MC_CPU1") and locking
of "MCU1" is enough.
However, locking of the communication channel is the correct thing to do.