ethercatmcCabinet.template: remove hardcoded parameter name
The PV that receives the 32 bit value of the cabinet status had
"Cabinet" as hardcoded asyn parameter name, whereas the
corresponding parameter creator in ethercatmcCabinet.iocsh
loads its name as $(R). This lead to the parameter not being
found unless the user loads R as "Cabinet", therefore fix was
needed. For consistency with the others, this commit
substitutes the hardcoded parameter name by $(R). It also adds
descriptions to the bits PVs, which adds extra information
about their meaning.

This closes #15 and closes #21.