ethercatmcCabinet: hardcode cabinet paramName in template and iocsh
In PR #22 the cabinet status asyn parameter name was changed to
be defined by the macro $R, but this generates a problem because
the PILS device for this is defined as "Cabinet#0", which imposes
the parameter name to be "Cabinet". This commits hardcodes the
parameter name properly, so that it cannot be user-defined.