op/Boy/tools/py: Use raw strings for regular expressions
When using regular expressions ("re.compile()") it is good
practice to use "raw strings" when special characters are used.
Special in the sence of the regexp.
As a string, the backslash is special. Forwarding a backslash
into the regexp needs to backslashes - or raw strings.

This was noticed by python 3.13, see
https://docs.python.org/3/library/re.html#module-re