Expose auto-power mode in production code
It happened quite often that a user doing an experiment tries to move a
motor that is powered off and needs to be enabled. This is because the
motion control engineers have chosen to keep the axis disabled and not
apply power to the motor. Neither have they chosen to implement an
auto-power-on feature inside the motion controller itself.
As a workaround, the auto-power-on feature inside the EPICS driver
has been enabled by default.
The disadvantage is that the auto-power on may take up to 6 seconds to
complete, and under this time the communication to all other axes is
blocked. As of today, the motor will be powered on once, and never
powered off from EPICS.

As the end-users do not communicate with the EPICS IOC directly, but use
a high-level control system like NICOS, the information about
powered-off-motors that are powered on in EPICS needs to be displayed to
the end user.
The power information is in the CNEN field of the motorRecord. The
auto-power feature can be found in an additional record.

Move 'record(longout, "$(P)$(R)-PwrAuto")' from ethercatmcdebug.template
into ethercatmc.template so that it is available in production
environment.