Clean up printing of changed reason bits
There are 24 aux bits in the status word. Not all of may be used.
The used ones are defined in the flags, retreived in initial
reading of the indexer.
When any of the bits are changed (or more than one) there is a printout
in the IOC log.
In the same status word there are 4 reason bits defined.
Limit switches are indicated here. Other alarms may use them as well
on non-motor devices.
To make a long story short:
When printing out the changed bits, 24 aux bits + 4 reason bits
have been printed since a long time.
Assuming that everything is a motor, parameters for NamBit27..NamBit24
had been defined. However, devices like Cabinet don't use those at all.
Remove all the special handling of bit 24..27.
Cleanup and rename variables to make more clear what is going on.

Changes to be committed:
    modified:   ethercatmcApp/src/ethercatmcController.cpp
    modified:   ethercatmcApp/src/ethercatmcController.h
    modified:   ethercatmcApp/src/ethercatmcIndexer.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerV2.cpp