ethercatmcController.cpp: look at interruptAccept again
The big cleanup removing non-used code did remove the
interruptAccept check in the poller.
Basically commit fad8129c
    ethercatmcController.cpp: Take care about interruptAccept

was unintentionally removed.
The effect was that the definition for mbbi records was lost,
ethercatmcIndexerAxis::readEnumsAndValueAndCallbackIntoMbbi()
did not have any effect.
Add a debug print in ethercatmcIndexerAxis.cpp

Solution:
Restore the handling of interruptAccept in the poller.

Changes to be committed:
    modified:   ethercatmcApp/src/ethercatmcController.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp