ethercatmcShutterV2: More NamAuxBit
The record names of the Aux Bit Names consistent with motors:
NamAuxBit is used for the suffix.
However, there is one difference:
The macro in the template file is
"$(P)$(R)-NamAuxBit0"
(upstream EPICS could have used "$(P)$(R):NamAuxBit0", but that is not
 according to the naming convention at ESS).

For the shutter we ommit the '-' and use "$(P)$(R)NamAuxBit0".
This gives names like XXX:ShtNamAuxBit0, which is readable as well.
(For the motors we keep the '-' to avoid breaking existing installations,
and this needs to be done on a different branch anyway)

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmcShutterV2.template
    modified:   ethercatmcExApp/op/Boy/IOC.opi
    modified:   ethercatmcExApp/op/Boy/tools/IOC-Sht.mid
    modified:   iocsh/ethercatmcShutterV2.iocsh