Reset the -ErrRst PV to 0 when reset is done
The "-ErrRst" PV triggers a reset command inside the motion controller.
Once a "1" is written to the record, it stays at "1" forever.
However, when "1" is written again, a new reset is send to
the controller. (So the reset works as expected).

Improve the behaviour of the record:
Reset it to "0" once the reset has been done inside the controller.

While there, integrate the one-line-method resetAxis() into the code.

Changes to be committed:
    modified:   ethercatmcApp/Db/ethercatmc.template
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.h