ethercatmcIndexerAxis: ErrTxt for Error with reason bits
The errTxt and the msgTxt provide useful information.
The pneumatic shutters have a new behaviour:
Error is set with errorId == 0.
Handle this better and show the reason bits.
If more than one reason bit is set, show the hex value.

Especially: avoid the confusing "TwinCAT ERR 0".
Changes to be committed:
    modified:   .../ethercatmcIndexerAxis.cpp
    modified:   pytests36/941_Simulator-MsgTxt-ErrTxt.py