ethercatmcIndexer: Support for an 1E0C device
Support for "discrete output with status word" (or so) devices.
The 1E0C is similar to the 1E04.
The main reason is to get the errorID from the controller into EPICS.
Bit 0 of the reserved field is used as 1 additional bit for the numeric
ErrorID, extending the range from 0x0000..0xFFFF to 0x0000..0x1FFFF.
All errorID below 0xFFFF are owned by the vendor of TwinCAT
and all errorID are owned by us.

Changes to be committed:
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerV2.cpp