Actuators and pneumatics error handling
Make it possible to forward errror IDs for devices that
are not motors:
Different kind of actuators, pneumatic shutters, insertable beam stops
or others.
In general, add support for the 1E0C device which has an errorID.
Define "home made errror IDs" in the range 0x10000..0x1FFFF and predefine
some for our "real live shutters".

Split the ethercatmcShutter.template into smaller templated and make
the integration of "not controlable, read only" shutters easier.

Add the new device into the simulator and add a test case
to verify the new error handling.

Changes to be committed:
    modified:   ethercatmcApp/Db/Makefile
    renamed:    ethercatmcApp/Db/ethercatmcShutter.template -> ethercatmcApp/Db/ethercatmcActuator.template
    new file:   ethercatmcApp/Db/ethercatmcActuatorActivate.template
    new file:   ethercatmcApp/Db/ethercatmcOpenedClosed.template
    modified:   ethercatmcApp/src/ethercatmcController.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerV2.cpp
    new file:   iocsh/ethercatmcActuator.iocsh
    modified:   iocsh/ethercatmcShutter.iocsh
    new file:   test/pytests36/943_Simulator_Pneumatics.py
    modified:   test/pytests36/Makefile
    modified:   test/simulator/indexer.c
    modified:   test/startup/st.sim-indexer.iocsh