t/.../943: Add test for AuxBits07
Add a test for the strings in the AuxBits07 record.
This test should have been here for a longer time.
Especially as it would had catched the regression fixed in
  commit d4666240da
      ethercatmcController.cpp: look at interruptAccept again

This very commit makes sure that the mbbi record gets it value
via the callback after reading the text strings from the MCU:

- Initalize all records.
- Records (e.g. mbbi) set up callback to the driver
- The driver communicates to the MCU
- callbacks are called