ethercatmcIndexer/indexerParamWrite(): Avoid endless loop
When changing JVEL while already jogging, a new moveVelocity()
is commanded.
However, this is handled really badly with older motor versions when auto-power-on
is used.
Trying to enable the motor leads to a endless loop.

Solution:
Count the counter always to leave the loop.
As an other improvement simply ignore the power-on while moving,
the motor should be powered on.