test...910: Fix timing problem
910_Simulator_JOG_BDST_RTRY_FRAC did sometimes fail:
The motor starts a JOG, and then is stopped by resetting the JOG[FR]
field.
Once stopped, add a predifined position, backlash should be done.
The problem is, that forcing the simulator into a position
did stop the movement.
When this stopped is reported to the motorRecord too early, before setting
JOG[FR] to 0, the motorRecord thinks that the motor had stopped with being
told to stop.
This stopped is treated as unexpected, and the backlash will not be done.

Solution:
Remove the blind sleep and use proper waitForValueChanged().
Use bManualSimulatorMode=1 to simulate a running motor, until it
is told to be stopped.