Better handling of flickering encoder when not busy
A high-resolution encoder may give us a new postion in every poll().
The current code uses the old- and new position to update the direction
bit, which now goes force and back, with every (?) poll.
This results in a change of the MSTA field in the motorRecord.
And the TDIR field as well.
Looking at TDIR:
It is definded as "travelling direction".
Note:
The changed position can be filtered out with help of ADEL/MDEL.

Solution:
Update direction only when the motor is not activly moving.
While there, clean the code up.
Additionally, set the moving bit first when the first change of
actual position is detected.
Reset moving when the controller is not busy any more.

Changes to be committed:
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.h