Improve the parameter interface
Refactor duplicated code in the parameter interface for reading
and writing.

Improve the retry-logic: instead an increasing timeout,
use always a 5 msec sleep between the retries.
This should give a good performance when the PLC cycle is 10msec.

In order to handle setups with many axes better, limit the total time
for the parameter interface to 300msec.

Changes to be committed:
    modified:   ethercatmcApp/src/ethercatmcController.h
    modified:   ethercatmcApp/src/ethercatmcIndexer.cpp
    modified:   ethercatmcApp/src/ethercatmcIndexerAxis.h