simulator: Keep param interface busy while JOG/HOM
The simulator behaves different from what the real MCU currently does:
When a home- or moveVelocity function is started,
the param interface may stay busy the whole time.
Simulate this here.

Note:
In the long run it makes sense to return to then non-busy
parameter handling. Otherwise changing of the velocity during
moveVelocity will be blocked.
In other words: This could be made configurable in the simulator
and test cases should be written in future commits.