ethercatmcIndexerAxis: Allow JOG and JVEL change
Some controllers do allow the change of the velocity while
moving with constant velocity.
This can be triggered by writing to the JOGF/JOGR field in the
motorRecord and, while the motor is moving, write a new value
to theJVEL field.
The generic driver will call moveVelocity() again, with a new value
for maxVelocity.

Allow this combination,
when we know that that the last call was a moveVelocity.