Add a generic analog output template
The current template named
ethercatmcAnalogOutputReadback.template
is used for most analog output records.
However, before we change it, create a new template that allows all
fields to be configured, even the info:READBACK ones.
Put a iocsh file around it, which uses all existing parameters.
More may be added to the template, but
ethercatmcAnalogOutputGenericV1.iocsh should be kept as is.
This will be used in the next commit.

Changes to be committed:
    modified:   ethercatmcApp/Db/Makefile
    new file:   ethercatmcApp/Db/ethercatmcAnalogOutputGeneric.template
    new file:   iocsh/ethercatmcAnalogOutputGenericV1.iocsh